requirements.txt
tweepy
requests
mysql-connector-python